refactor: Remove unused imports and code
This commit removes unused imports and code from various files in the codebase. The changes include removing the unused import of `ContactFormEmail` in `_actions.ts`, removing the unused import of `v` and `Infer` in `schema.ts`, removing the unused import of `Authenticated` and `Spinner` in `convex-provider.tsx`, removing the unused imports of `Image` and `Link` in `RentPlaceFeaturedCarousel.tsx`, removing the unused import of `v` in `contact.ts`, and removing the unused import of `CustomModalCalendly` in `[slug]/page.tsx`. These changes help to clean up the code and improve maintainability.